From Andrew - a number of changes for MSVC. cst.c: whack dead variable. csv_util.c, globals.c, waypt.c : dont explictly init global. gdb.c, msroute.c: return after 'fatal'. (decoration...) gpx.c: fix cast. ignrando.c, nmn4.c: cast _count to int. waypt.c: init 'no' before use. cet/* Zero length arrays aren't C89. msvc/GPSBabel.vcproj: Update.

+++ trunk/gpsbabel/ignrando.c
@@ -276,7 +276,7 @@ ignr_write(void)
 	if (index_opt != NULL)
 	{
 		track_index = atoi(index_opt);
-		if ((track_index < 1) || (track_index > track_count()))
+		if ((track_index < 1) || (track_index > (int) track_count()))
 		    fatal(MYNAME ": Invalid track index %d (we have currently %d track(s))!\n", 
 			track_index, track_count());
 	}
